avoid some compiler warnings on solaris-2.9/sparc (gcc-3.3.5)

+++ src/hid/lesstif/main.c
@@ -432,7 +432,7 @@ command_parse (char *s)
 
   for (cp = s; *cp; cp++)
     {
-      if (isspace (*cp))
+      if (isspace ((int) *cp))
 	ws = 1;
       else
 	{
@@ -446,7 +446,7 @@ command_parse (char *s)
   ws = 1;
   for (cp = s; *cp; cp++)
     {
-      if (isspace (*cp))
+      if (isspace ((int) *cp))
 	{
 	  ws = 1;
 	  *cp = 0;
